Add Chinese (Traditional)
It is necessary to rename from lang = "zh" into "zh_CN" before the new strings lang = "zh_TW" can be added.  Otherwise, the Summary and Description will always display Simplified Chinese even in Kodi even when the default font is set to Traditional Chinese.  That is not a comfortable look on a system defaulted to Chinese Traditional.
Moreover, the characters, the grammar, usage and culture are very difference.